Decode ratchetTree extension
